requirements.txt
torch
torchvision
pyyaml
pydantic
pandas
matplotlib
seaborn
jupyterlab
